gemini/tuning/genai-mlops-tune-and-eval/requirements.txt
annotated-types==0.7.0
<<<<<<< HEAD
cachetools==5.5.1
=======
cachetools==5.5.2
>>>>>>> main
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
docstring_parser==0.16
<<<<<<< HEAD
google-api-core==2.24.0
google-auth==2.38.0
google-cloud-aiplatform==1.78.0
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.14.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.14.0
grpcio==1.70.0
grpcio-status==1.62.3
idna==3.10
kfp==2.11.0
kfp-pipeline-spec==0.6.0
kfp-server-api==2.3.0
kubernetes==30.1.0
numpy==2.2.2
oauthlib==3.2.2
packaging==24.2
proto-plus==1.25.0
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
=======
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-aiplatform==1.87.0
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
idna==3.10
kfp==2.12.1
kfp-pipeline-spec==0.6.0
kfp-server-api==2.4.0
kubernetes==30.1.0
numpy==2.2.4
oauthlib==3.2.2
packaging==24.2
proto-plus==1.26.1
protobuf==4.25.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.1
pydantic_core==2.33.1
>>>>>>> main
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==0.10.1
rsa==4.9
<<<<<<< HEAD
shapely==2.0.6
six==1.17.0
tabulate==0.9.0
typing_extensions==4.12.2
=======
shapely==2.0.7
six==1.17.0
tabulate==0.9.0
typing_extensions==4.13.0
>>>>>>> main
urllib3==1.26.20
websocket-client==1.8.0
